requirements.txt
numpy
matplotlib
pytorch>=1.7.0
torchvision
scikit-learn
